Re-add execution of doctests
Follow up to https://github.com/crate/crate-python/pull/767
